Model updates.
Added export_id for volumes (needed for NFS) and made it unique within the cluster.

Added access_level choice for rw/ro to shares.

Added access prefixes and IPs for shares and removed them from the mount, since they are more about access control than where and how it should be mounted.

Updated the custom_field_data JSON encoder to the default for the latest version of NetBox.